Virtualize local variables
Summary:
Instead of storing local python variables in the memory provided
by PyFrameObject, locals are loaded/stored from alloca'd
memory, which will be translated by the backend into either
physical registers or machine stack slots.

When virtual locals are enabled, the signature of the generated
function is also modified so that Python arguments are passed as
native function arguments. This, in conjunction with speculative
devirtualization, should significantly improve the performance
of function calls.

There is a significant limitation of the current implementation:
frames with virtual locals cannot be inspected using
PyFrame_FastToLocals() or the "f_locals" property of the frame
object, since these assume the locals are kept inside the PyFrameObject.
This breaks many tests, and functions which use super(). To get
around this, this change hard-codes a list of hashes of code objects
which cannot use virtual locals. This solution is fragile and should be
temporary. A better solution is to implement PyFrame_FastToLocals()
by examining the machine stack.