Implement two algorithms for finding the dominator tree
Summary:
Implements the SEMI-NCA algorithm, which is based on the Lengauer-Tarjan algorithm.
This is same algorithm LLVM uses, and the one that will be used normally.

The second is an optimized version of the iterative dataflow algorithm. It
is reasonably fast, but with worse asymptotic performance than SEMI-NCA.
It is used for verification purposes. When run with PYJITDEBUG flag set to
2 or higher, both algorithms are run on every function, and their output
is checked for equality.