Add light-weight frames
Summary:
Add "PyRunFrame" structure, which will eventually become the source of truth for JIT function call frames.

In this change, both a PyRunFrame and PyFrameObject are created for every frame
(both interpreter and JIT frames). The next step is to make it unnecessary for JIT
frames to instantiate a PyFrameObject, instead relying on the PyRunFrame only.

Any PyRunFrame can be materialized into a regular PyFrameObject. This happens whenever a genuine
PyFrameObject is needed (such as for tracebacks).