Remove dependency on PyFrameObject for (most) JIT-generated code
Summary:
This change introduces a flag in the JIT generator, "use_frame_object", which controls
whether the JIT-generated code uses a PyFrameObject, or whether its local state is
completely virtual. (i.e. exists in registers or on the thread stack)

As a result of this change, the JIT now produces PyRunFrame's which are unmaterialized
(do not have a corresponding PyFrameObject) by default. When a running frame is
materialized, the corresponding frame object may exist in a partial state where the
source of truth for the runtime state is actually in the PyRunFrame (or on the thread stack).
To handle this correctly, this change also adds getters for PyFrameObject fields (f_lasti,
f_blockstack, and f_localsplus) which may need to be pulled from another location.