LLVM struct size fix
Summary:
Allocating a PyRunFrame struct on the stack requires knowing sizeof(PyRunFrame),
but the LLVM backend uses an opaque structure which does not have this information.

This changes the LLVM backend so that the bitcode struct types have size information.

Also fixes a stack alignment issue in the trampoline, which was causing an internal
LLVM compiler crash. (unaligned floating point mov)